fix: NexusRESTClient circular import
